The start of Parameterised reports.
Removing all generate functions for TreeView elements into their respective classes, in order to not have massive switch statements for deciding which element is of what type. So the generating can then be generalised.

The generation happens through a while loop which will be deleting elements as they are added to the XML document, and also recursively going through each generateXML function as before.