Creating the generateXML functions for respective classes. Some functions are null to fix compilation errors. Others have started to be implemented such as the Teams, Projects and People XML generation.
In order to traverse the checked Items I have made it so that each element will iterate through it's children and see if it is a checked item, this is in order to maintain hierarchy and the previous format of XML. When the child is not available in the case of Allocations and Teams on projects it needed to be called manually. The same thing needs to be done for skills on people.

Also fixed a bug within the TeamEditScene that was not assigning roles to people, it was just dealing with roles on the Teams side of things. When the TeamEditScene was rewritten, a new role was created called NONE, this makes more logical sense than OTHER but the OTHER role was not replaced, OTHER has now been replaced by NONE.