Unit tests for XML generation of Allocations, Stories and Roles.
Following the same logic of Skills unit test with only checking that the text exists and is in the correct order. Essentially copied and pasted the code for the above functions with some minor changes to make sure it works for each one. Also updated the Javadoc in the Skills test.