Created tests for Person and Person with skills, changed person to have skills generated within a skills element of person as opposed to seperately.
Created two different tests in Person, one for having skills and one without skills to see if the hierarchy is generated how it should be. Also added in a check to see how many elements exist directly under the returned elements of each of the current tests so far, as to a way to check structure in a way.

When checking skills under people I did not check the details within skills as that is done under the unit tests for skills.

 Also added in a skills tag in person generateXML as it was not the way I expected it to be printed out. Thanks tests :)