Started converting the edit scenes from Panes to Tabs.
Currently I have done Skill, Team and Workspace. Created new Classes to hold the edit tab information and marked the old edit scenes as Deprecated.

Also introduced another constructor to the Element scene (ie TeamScene, SkillScene) where an edit boolean is taken in, and it shows an edit tab. This means we do not need to create a seperate scene just for the edit tab.