Changed custom combo box so it can take in a list of items. This fixed the combo box issues the team edit scene was having (Migrating from ComboBox to CustomComboBox lost some functionality).
Fixed an exception on the PersonInfoTab where a null team was not being correctly checked. Instead, the unassigned team is now being checked.

Changed the "none" role to "" for consistency throughout the project.

Removed the project reference from team (was never being set) and removed the setProject method. Adjusted the getProject method so it returns the teams current project by getting the current allocations project.